requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
requests==2.31.0
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.3.7
cryptography==41.0.4
urllib3==2.0.5
